Convert miliseconds to int

+++ models/unit_type_parsers/seconds_parser.py
@@ -5,5 +5,5 @@ class SecondsParser(BaseParser):
     def __init__(self):
         super(SecondsParser, self).__init__('S')
 
-    def convert_to_bytes(self, value: int) -> bytes:
-        return (value * 60).to_bytes(2, 'big')
+    def convert_to_bytes(self, value: float) -> bytes:
+        return int((value * 60)).to_bytes(2, 'big')
